ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.25.2
asyncpg==0.29.0
python-multipart==0.0.6
joblib==1.3.2 